add bash_profile
A slight amount of refactoring was needed before being able to push to a
public place. There are still some unnecessary lines in here and others
that I don't understand due to copy-paste years ago. I hope to figure
those out or get rid of them as soon as possible.