Create README.md
The README was lost in my initial push. This is because I did a setup based on a bare repo and a custom bash alias (this https://developer.atlassian.com/blog/2016/02/best-way-to-store-dotfiles-git-bare-repo/) and messed up pulling the changes from creating it on the web interface. In other words, I lost my 'Initial commit' which is not a big deal.