preserve virtualenv prepending to prompt
Using __git_ps1 from PROMPT_COMMAND is quite convenient and looks very
simple, but it overrides anything else that wants to modify the prompt.
This fix only solves the case where activating a virtualenv prepends the
name of the virtualenv to the prompt. Fortunately, this is the only
other case I have for now.